Display percent gradient instead of just current elevation
The points get replayed as sections (where a section has a start point and end point from consecutive points in the gpx data). Then display the gradient of the current section (instead of elevation of the current point).

+++ TurboTrainer.Tests/GpxPointTestExtensions.cs
@@ -7,7 +7,7 @@ namespace TurboTrainer.Tests
 {
     public static class GpxPointTestExtensions
     {
-        public static void AssertPoint(this GpxPoint point, double expectedLatitude, double expectedLongitude, double expectedElevation, string expectedTime)
+        public static void AssertPoint(this GpxPoint point, decimal expectedLatitude, decimal expectedLongitude, decimal expectedElevation, string expectedTime)
         {
             Assert.That(point.Latitude, Is.EqualTo(expectedLatitude));
             Assert.That(point.Longitude, Is.EqualTo(expectedLongitude));
